+++ opentelemetry-instrumentation/src/opentelemetry/instrumentation/auto_instrumentation/sitecustomize.py
@@ -15,7 +15,6 @@
 from logging import getLogger
 from os import environ
 from os.path import abspath, dirname, pathsep
-from re import sub
 
 from pkg_resources import iter_entry_points
 
@@ -26,6 +25,7 @@ from opentelemetry.instrumentation.distro import BaseDistro, DefaultDistro
 from opentelemetry.instrumentation.environment_variables import (
     OTEL_PYTHON_DISABLED_INSTRUMENTATIONS,
 )
+from opentelemetry.instrumentation.utils import _python_path_without_directory
 from opentelemetry.instrumentation.version import __version__
 
 logger = getLogger(__name__)
@@ -111,10 +111,8 @@ def _load_configurators():
 
 def initialize():
     # prevents auto-instrumentation of subprocesses if code execs another python process
-    environ["PYTHONPATH"] = sub(
-        rf"{dirname(abspath(__file__))}{pathsep}?",
-        "",
-        environ["PYTHONPATH"],
+    environ["PYTHONPATH"] = _python_path_without_directory(
+        environ["PYTHONPATH"], dirname(abspath(__file__)), pathsep
     )
 
     try: